int overflow

+++ asm_src/includes/asm.h
@@ -6,7 +6,7 @@
 /*   By: nsheev <nsheev@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/12/11 17:41:33 by swedde            #+#    #+#             */
-/*   Updated: 2020/01/17 17:09:57 by nsheev           ###   ########.fr       */
+/*   Updated: 2020/01/22 16:38:05 by nsheev           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -51,7 +51,7 @@ typedef struct		s_token
 
 union				u_int
 {
-	unsigned long	a;
+	unsigned		a;
 	struct			s_ch
 	{
 		unsigned char	c1:8;
@@ -88,7 +88,7 @@ typedef struct		s_all
 }					t_all;
 
 t_all				*set_def_gen(char *s);
-int					write_int_to_file(t_all	*gen, unsigned long int to_file);
+int					write_int_to_file(t_all	*gen, unsigned int to_file);
 void				do_exit(char **error_mes, t_all *gen);
 int					reading(t_all *gen);
 char				*char_to_string(char c);
@@ -108,6 +108,6 @@ int					write_byte_to_file(t_all *gen, unsigned char c);
 int					write_nam_com_to_file(t_all *gen, int len, int type);
 void				print_token_type(int i);
 
-void		print_token(t_token*	token);
+void				print_token(t_token	*token);
 
 #endif
